Fixed module paths after refactor to src format
