+++ scripts/debug_model_state.py
@@ -1,5 +1,5 @@
-from src.monarchs.core import reload_state
-from src.monarchs.core import get_2d_grid
+from monarchs.core import reload_state
+from monarchs.core import get_2d_grid
 from matplotlib import pyplot as plt
 
 path = 'C:/Users/jdels/Documents/Work/MONARCHS_runs/ARCHER2_030225/dump.nc'
