+++ src/monarchs/physics/Numba/heateqn_nb.py
@@ -14,12 +14,12 @@ import numpy as np
 from numba import cfunc
 from NumbaMinpack import minpack_sig
 
-from src.monarchs.physics.Numba.extract_args import (
+from monarchs.physics.Numba.extract_args import (
     extract_args,
     extract_args_fixedsfc,
     extract_args_lid,
 )
-from src.monarchs.physics.surface_fluxes import sfc_flux
+from monarchs.physics.surface_fluxes import sfc_flux
