+++ tests/serial/create_test_IceShelf.py
@@ -6,8 +6,8 @@ Create an instance of an IceShelf object for testing with.
 
 
 def frozen_testcase():
-    from src.monarchs.core.iceshelf_class import IceShelf
-    from src.monarchs.core.initial_conditions import rho_init_emp
+    from monarchs.core.iceshelf_class import IceShelf
+    from monarchs.core.initial_conditions import rho_init_emp
 
     firn_depth = 39.99973571
     vert_grid = 400
